requirements.txt.txt
Flask==2.3.2
Flask-Cors==3.0.10
llama-index==0.1.0
python-dotenv==1.0.0
